♻️ Refactor: 메인 페이지 query string 갱신 로직 분리
